+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/SolutionI.java
@@ -67,11 +67,6 @@ public class SolutionI implements Solution {
     }
 
     @java.lang.SuppressWarnings("all")
-    public Solution toSolution() {
-        return this.problem.toSolution();
-    }
-
-    @java.lang.SuppressWarnings("all")
     public Solution asSolution() {
         return this.problem.asSolution();
     }